# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Kiali project, observability for the Istio service mesh",
    "title": "Kiali",
    "version": "_"
  },
  "basePath": "/api",
  "paths": {
...
    "/namespaces": {
      "get": {
        "description": "Endpoint to get the list of the available namespaces",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "namespaces"
        ],
        "operationId": "namespaceList",
        "responses": {
          "200": {
            "$ref": "#/responses/namespaceList"
          },
          "500": {
            "$ref": "#/responses/internalError"
          }
        }
      }
    },
    "/namespaces/graph": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "graphs"
        ],
        "summary": "The backing JSON for a namespaces graph.",
        "operationId": "graphNamespaces",
        "parameters": [
          {
            "type": "string",
            "default": "run all appenders",
            "x-go-name": "Name",
            "description": "Comma-separated list of Appenders to run. Available appenders: [aggregateNode, deadNode, healthConfig, idleNode, istio, responseTime, securityPolicy, serviceEntry, sidecarsCheck, throughput].",
            "name": "appenders",
            "in": "query"
          },
          {
            "type": "string",
            "default": "none",
            "x-go-name": "Name",
            "description": "Comma-separated list of desired node boxing. Available boxings: [app, cluster, namespace, none].",
            "name": "boxBy",
            "in": "query"
          },
          {
            "type": "string",
            "default": "10m",
            "x-go-name": "Name",
            "description": "Query time-range duration (Golang string duration).",
            "name": "duration",
            "in": "query"
          },
          {
            "type": "string",
            "default": "workload",
            "x-go-name": "Name",
            "description": "Graph type. Available graph types: [app, service, versionedApp, workload].",
            "name": "graphType",
            "in": "query"
          },
          {
            "type": "string",
            "default": "false",
            "x-go-name": "Name",
            "description": "Flag for including edges that have no request traffic for the time period.",
            "name": "includeIdleEdges",
            "in": "query"
          },
          {
            "type": "string",
            "default": "false",
            "x-go-name": "Name",
            "description": "Flag for injecting the requested service node between source and destination nodes.",
            "name": "injectServiceNodes",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "Name",
            "description": "Comma-separated list of namespaces to include in the graph. The namespaces must be accessible to the client.",
            "name": "namespaces",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "default": "now",
            "x-go-name": "Name",
            "description": "Unix time (seconds) for query such that time range is [queryTime-duration..queryTime]. Default is now.",
            "name": "queryTime",
            "in": "query"
          },
          {
            "type": "string",
            "default": "requests",
            "x-go-name": "Name",
            "description": "How to calculate gRPC traffic rate. One of: none | received (i.e. response_messages) | requests | sent (i.e. request_messages) | total (i.e. sent+received).",
            "name": "rateGrpc",
            "in": "query"
          },
          {
            "type": "string",
            "default": "requests",
            "x-go-name": "Name",
            "description": "How to calculate HTTP traffic rate. One of: none | requests.",
            "name": "rateHttp",
            "in": "query"
          },
          {
            "type": "string",
            "default": "sent",
            "x-go-name": "Name",
            "description": "How to calculate TCP traffic rate. One of: none | received (i.e. received_bytes) | sent (i.e. sent_bytes) | total (i.e. sent+received).",
            "name": "rateTcp",
            "in": "query"
          },
          {
            "type": "string",
            "default": "95",
            "x-go-name": "Name",
            "description": "Used only with responseTime appender. One of: avg | 50 | 95 | 99.",
            "name": "responseTime",
            "in": "query"
          },
          {
            "type": "string",
            "default": "request",
            "x-go-name": "Name",
            "description": "Used only with throughput appender. One of: request | response.",
            "name": "throughput",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/graphResponse"
          },
          "400": {
            "$ref": "#/responses/badRequestError"
          },
          "500": {
            "$ref": "#/responses/internalError"
          }
        }
      }
    },
    "/namespaces/{namespace}": {
      "patch": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "namespaces"
        ],
        "summary": "Endpoint to update the Namespace configuration using Json Merge Patch strategy.",
        "operationId": "namespaceUpdate",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Name",
            "description": "The namespace name.",
            "name": "namespace",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/namespaceResponse"
          },
          "400": {
            "$ref": "#/responses/badRequestError"
          },
          "404": {
            "$ref": "#/responses/notFoundError"
          },
          "500": {
            "$ref": "#/responses/internalError"
          }
        }
      }
    },
...
    "/namespaces/{namespace}/services/{service}/graph": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "graphs"
        ],
        "summary": "The backing JSON for a service node detail graph.",
        "operationId": "graphService",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Name",
            "description": "The cluster name. If not supplied queries/results will not be constrained by cluster.",
            "name": "container",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "Name",
            "description": "The namespace name.",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "Name",
            "description": "The service name.",
            "name": "service",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "default": "run all appenders",
            "x-go-name": "Name",
            "description": "Comma-separated list of Appenders to run. Available appenders: [aggregateNode, deadNode, healthConfig, idleNode, istio, responseTime, securityPolicy, serviceEntry, sidecarsCheck, throughput].",
            "name": "appenders",
            "in": "query"
          },
          {
            "type": "string",
            "default": "none",
            "x-go-name": "Name",
            "description": "Comma-separated list of desired node boxing. Available boxings: [app, cluster, namespace, none].",
            "name": "boxBy",
            "in": "query"
          },
          {
            "type": "string",
            "default": "10m",
            "x-go-name": "Name",
            "description": "Query time-range duration (Golang string duration).",
            "name": "duration",
            "in": "query"
          },
          {
            "type": "string",
            "default": "workload",
            "x-go-name": "Name",
            "description": "Graph type. Available graph types: [app, service, versionedApp, workload].",
            "name": "graphType",
            "in": "query"
          },
          {
            "type": "string",
            "default": "now",
            "x-go-name": "Name",
            "description": "Unix time (seconds) for query such that time range is [queryTime-duration..queryTime]. Default is now.",
            "name": "queryTime",
            "in": "query"
          },
          {
            "type": "string",
            "default": "requests",
            "x-go-name": "Name",
            "description": "How to calculate gRPC traffic rate. One of: none | received (i.e. response_messages) | requests | sent (i.e. request_messages) | total (i.e. sent+received).",
            "name": "rateGrpc",
            "in": "query"
          },
          {
            "type": "string",
            "default": "requests",
            "x-go-name": "Name",
            "description": "How to calculate HTTP traffic rate. One of: none | requests.",
            "name": "rateHttp",
            "in": "query"
          },
          {
            "type": "string",
            "default": "sent",
            "x-go-name": "Name",
            "description": "How to calculate TCP traffic rate. One of: none | received (i.e. received_bytes) | sent (i.e. sent_bytes) | total (i.e. sent+received).",
            "name": "rateTcp",
            "in": "query"
          },
          {
            "type": "string",
            "default": "95",
            "x-go-name": "Name",
            "description": "Used only with responseTime appender. One of: avg | 50 | 95 | 99.",
            "name": "responseTime",
            "in": "query"
          },
          {
            "type": "string",
            "default": "request",
            "x-go-name": "Name",
            "description": "Used only with throughput appender. One of: request | response.",
            "name": "throughput",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/graphResponse"
          },
          "400": {
            "$ref": "#/responses/badRequestError"
          },
          "500": {
            "$ref": "#/responses/internalError"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Config": {
      "type": "object",
      "properties": {
        "duration": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Duration"
        },
        "elements": {
          "$ref": "#/definitions/Elements"
        },
        "graphType": {
          "type": "string",
          "x-go-name": "GraphType"
        },
        "timestamp": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Timestamp"
        }
      },
      "x-go-package": "github.com/kiali/kiali/graph/config/cytoscape"
    },
...
    "EdgeData": {
      "type": "object",
      "properties": {
        "destPrincipal": {
          "description": "App Fields (not required by Cytoscape)",
          "type": "string",
          "x-go-name": "DestPrincipal"
        },
        "id": {
          "description": "Cytoscape Fields",
          "type": "string",
          "x-go-name": "ID"
        },
        "isMTLS": {
          "type": "string",
          "x-go-name": "IsMTLS"
        },
        "responseTime": {
          "type": "string",
          "x-go-name": "ResponseTime"
        },
        "source": {
          "type": "string",
          "x-go-name": "Source"
        },
        "sourcePrincipal": {
          "type": "string",
          "x-go-name": "SourcePrincipal"
        },
        "target": {
          "type": "string",
          "x-go-name": "Target"
        },
        "throughput": {
          "type": "string",
          "x-go-name": "Throughput"
        },
        "traffic": {
          "$ref": "#/definitions/ProtocolTraffic"
        }
      },
      "x-go-package": "github.com/kiali/kiali/graph/config/cytoscape"
    },
    "EdgeWrapper": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/EdgeData"
        }
      },
      "x-go-package": "github.com/kiali/kiali/graph/config/cytoscape"
    },
    "Elements": {
      "type": "object",
      "properties": {
        "edges": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/EdgeWrapper"
          },
          "x-go-name": "Edges"
        },
        "nodes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/NodeWrapper"
          },
          "x-go-name": "Nodes"
        }
      },
      "x-go-package": "github.com/kiali/kiali/graph/config/cytoscape"
    },
...
    "GWInfo": {
      "description": "GWInfo contains the resolved gateway configuration if the node represents an Istio gateway",
      "type": "object",
      "properties": {
        "ingressInfo": {
          "$ref": "#/definitions/GWInfoIngress"
        }
      },
      "x-go-package": "github.com/kiali/kiali/graph/config/cytoscape"
    },
    "GWInfoIngress": {
      "description": "GWInfoIngress contains the resolved gateway configuration if the node represents an Istio ingress gateway",
      "type": "object",
      "properties": {
        "hostnames": {
          "description": "Hostnames is the list of hosts being served by the associated Istio gateways.",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Hostnames"
        }
      },
      "x-go-package": "github.com/kiali/kiali/graph/config/cytoscape"
    },
...
    "HealthConfig": {
      "description": "HealthConfig maps annotations information for health",
      "type": "object",
      "additionalProperties": {
        "type": "string"
      },
      "x-go-package": "github.com/kiali/kiali/graph/config/cytoscape"
    },
...
    "NodeData": {
      "type": "object",
      "properties": {
        "aggregate": {
          "type": "string",
          "x-go-name": "Aggregate"
        },
        "app": {
          "type": "string",
          "x-go-name": "App"
        },
        "cluster": {
          "type": "string",
          "x-go-name": "Cluster"
        },
        "destServices": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ServiceName"
          },
          "x-go-name": "DestServices"
        },
        "hasCB": {
          "type": "boolean",
          "x-go-name": "HasCB"
        },
        "hasFaultInjection": {
          "type": "boolean",
          "x-go-name": "HasFaultInjection"
        },
        "hasHealthConfig": {
          "$ref": "#/definitions/HealthConfig"
        },
        "hasMissingSC": {
          "type": "boolean",
          "x-go-name": "HasMissingSC"
        },
        "hasRequestRouting": {
          "type": "boolean",
          "x-go-name": "HasRequestRouting"
        },
        "hasRequestTimeout": {
          "type": "boolean",
          "x-go-name": "HasRequestTimeout"
        },
        "hasTCPTrafficShifting": {
          "type": "boolean",
          "x-go-name": "HasTCPTrafficShifting"
        },
        "hasTrafficShifting": {
          "type": "boolean",
          "x-go-name": "HasTrafficShifting"
        },
        "hasVS": {
          "$ref": "#/definitions/VSInfo"
        },
        "id": {
          "description": "Cytoscape Fields",
          "type": "string",
          "x-go-name": "ID"
        },
        "isBox": {
          "type": "string",
          "x-go-name": "IsBox"
        },
        "isDead": {
          "type": "boolean",
          "x-go-name": "IsDead"
        },
        "isGateway": {
          "$ref": "#/definitions/GWInfo"
        },
        "isIdle": {
          "type": "boolean",
          "x-go-name": "IsIdle"
        },
        "isInaccessible": {
          "type": "boolean",
          "x-go-name": "IsInaccessible"
        },
        "isOutside": {
          "type": "boolean",
          "x-go-name": "IsOutside"
        },
        "isRoot": {
          "type": "boolean",
          "x-go-name": "IsRoot"
        },
        "isServiceEntry": {
          "$ref": "#/definitions/SEInfo"
        },
        "namespace": {
          "type": "string",
          "x-go-name": "Namespace"
        },
        "nodeType": {
          "description": "App Fields (not required by Cytoscape)",
          "type": "string",
          "x-go-name": "NodeType"
        },
        "parent": {
          "type": "string",
          "x-go-name": "Parent"
        },
        "service": {
          "type": "string",
          "x-go-name": "Service"
        },
        "traffic": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProtocolTraffic"
          },
          "x-go-name": "Traffic"
        },
        "version": {
          "type": "string",
          "x-go-name": "Version"
        },
        "workload": {
          "type": "string",
          "x-go-name": "Workload"
        }
      },
      "x-go-package": "github.com/kiali/kiali/graph/config/cytoscape"
    },
    "NodeWrapper": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/NodeData"
        }
      },
      "x-go-package": "github.com/kiali/kiali/graph/config/cytoscape"
    },
...
    "ProtocolTraffic": {
      "description": "ProtocolTraffic supplies all of the traffic information for a single protocol",
      "type": "object",
      "properties": {
        "protocol": {
          "type": "string",
          "x-go-name": "Protocol"
        },
        "rates": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Rates"
        },
        "responses": {
          "$ref": "#/definitions/Responses"
        }
      },
      "x-go-package": "github.com/kiali/kiali/graph/config/cytoscape"
    },
...
    "ResponseDetail": {
      "type": "object",
      "title": "ResponseDetail holds information broken down by response code.",
      "properties": {
        "flags": {
          "$ref": "#/definitions/ResponseFlags"
        },
        "hosts": {
          "$ref": "#/definitions/ResponseHosts"
        }
      },
      "x-go-package": "github.com/kiali/kiali/graph/config/cytoscape"
    },
    "ResponseFlags": {
      "description": "\"200\" : {\n\"-\"     : \"80.0\",\n\"DC\"    : \"10.0\",\n\"FI,FD\" : \"10.0\"\n}, ...",
      "type": "object",
      "title": "ResponseFlags is a map of maps. Each response code is broken down by responseFlags:percentageOfTraffic, e.g.:",
      "additionalProperties": {
        "type": "string"
      },
      "x-go-package": "github.com/kiali/kiali/graph/config/cytoscape"
    },
    "ResponseHosts": {
      "description": "\"200\" : {\n\"www.google.com\" : \"80.0\",\n\"www.yahoo.com\"  : \"20.0\"\n}, ...",
      "type": "object",
      "title": "ResponseHosts is a map of maps. Each response host is broken down by responseFlags:percentageOfTraffic, e.g.:",
      "additionalProperties": {
        "type": "string"
      },
      "x-go-package": "github.com/kiali/kiali/graph/config/cytoscape"
    },
    "Responses": {
      "description": "Responses maps responseCodes to detailed information for that code",
      "type": "object",
      "additionalProperties": {
        "$ref": "#/definitions/ResponseDetail"
      },
      "x-go-package": "github.com/kiali/kiali/graph/config/cytoscape"
    },
...
    "SEInfo": {
      "description": "SEInfo provides static information about the service entry",
      "type": "object",
      "properties": {
        "hosts": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Hosts"
        },
        "location": {
          "type": "string",
          "x-go-name": "Location"
        },
        "namespace": {
          "type": "string",
          "x-go-name": "Namespace"
        }
      },
      "x-go-package": "github.com/kiali/kiali/graph"
    },
...
    "ServiceName": {
      "type": "object",
      "properties": {
        "cluster": {
          "type": "string",
          "x-go-name": "Cluster"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "namespace": {
          "type": "string",
          "x-go-name": "Namespace"
        }
      },
      "x-go-package": "github.com/kiali/kiali/graph"
    },
...
    "VSInfo": {
      "type": "object",
      "title": "VSInfo contains the resolved VS configuration if the node has a VS attached.",
      "properties": {
        "hostnames": {
          "description": "Hostnames is the list of hostnames configured in the associated VSs",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Hostnames"
        }
      },
      "x-go-package": "github.com/kiali/kiali/graph/config/cytoscape"
    },
...
    "namespace": {
      "description": "A Namespace provide a scope for names\nThis type is used to describe a set of objects.",
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "annotations": {
          "description": "Specific annotations used in Kiali",
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Annotations"
        },
        "labels": {
          "description": "Labels for Namespace",
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Labels"
        },
        "name": {
          "description": "The id of the namespace.",
          "type": "string",
          "x-go-name": "Name",
          "example": "istio-system"
        }
      },
      "x-go-name": "Namespace",
      "x-go-package": "github.com/kiali/kiali/models"
    },
...
  },
  "responses": {
...
    "badRequestError": {
      "description": "BadRequestError: the client request is incorrect",
      "schema": {
        "type": "object",
        "properties": {
          "code": {
            "description": "HTTP status code",
            "type": "integer",
            "format": "int32",
            "default": 400,
            "x-go-name": "Code",
            "example": 400
          },
          "message": {
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    },
...
    "graphResponse": {
      "description": "HTTP status code 200 and cytoscapejs Config in data",
      "schema": {
        "$ref": "#/definitions/Config"
      }
    },
    "internalError": {
      "description": "A Internal is the error message that means something has gone wrong",
      "schema": {
        "type": "object",
        "properties": {
          "code": {
            "description": "HTTP status code",
            "type": "integer",
            "format": "int32",
            "default": 500,
            "x-go-name": "Code",
            "example": 500
          },
          "message": {
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    },
...
    "namespaceList": {
      "description": "List of Namespaces",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/namespace"
        }
      }
    },
    "namespaceResponse": {
      "description": "namespaceResponse is a basic namespace",
      "schema": {
        "$ref": "#/definitions/namespace"
      }
    },
...
    "notFoundError": {
      "description": "A NotFoundError is the error message that is generated when server could not find what was requested.",
      "schema": {
        "type": "object",
        "properties": {
          "code": {
            "description": "HTTP status code",
            "type": "integer",
            "format": "int32",
            "default": 404,
            "x-go-name": "Code",
            "example": 404
          },
          "message": {
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    },
...
  }
}